requirements.txt
streamlit
openai
streamlit-searchbox
fuzzywuzzy
boto3
geopy
pandas
pillow